fix(_math): Avoid forcecasting (#479)
The problem with the flag `py::array::forcecast` is that, e.g., in a
scenario where the scalars in the input array has a dtype which is
equivalent to `double` in C++, `pybind` will go through the signatures,
and if the first signature is specified with `float` and `forcecast` is
also specified, the `double`s will be converted to `float`s here,
instead of checking the next signature. This results in a decrease in
precision.

To resolve this, the `forcecast` flag is deleted. However, casting still
needs to be done for integer array arguments, therefore the functions
`numpy_to_matrix` and `numpy_to_vector` got extended with an additional
template argument `Flags`.