FIX(takagi): Scipy `sqrtm` behaviour change
**Problem**

In SciPy version 1.16, the `sqrtm` function has been rewritten, which caused a
the Takagi decomposition of Piquasso to yield wrong results in rare cases.

In the Takagi decomposition, the square root of a symmetric unitary matrix
(`Z`) is taken, and the code assumes that the `sqrtm` on such matrices results
in a unitary matrix. This is wrong, since nothing guarantees that a non-unitary
square root is returned, but practically, it was the case until SciPy version
1.16.

**Solution**

To avoid such errors, `sqrtm` has been replaced by using `eigh` directly to
obtain a square root which is unitary. For compatibility, the `eigh` and `inv`
functions are included as methods of `NumpyConnector` and
`TensorflowConnector`.

Sources:
- https://docs.scipy.org/doc/scipy/release/1.16.0-notes.html
- https://github.com/scipy/scipy/pull/22406
- https://journals.aps.org/pra/abstract/10.1103/PhysRevA.94.062109