chore(postselection): Remove `postselect_modes` input argument, `pq.Q`/`on_modes` to be used instead (#509)
Changes how modes are specified for postselection instructions.

**Context**

`PostSelectPhotons` and `ImperfectPostSelectPhotons` classes have the
input argument `postselect_modes` to specify which modes should be
postselected. However, all other instructions in Piquasso have a
different way to specify which modes the instruction should act on. This
mismatch is unfortunate because it's breaking Piquasso user interface
convention.

**Changes**

* Removes the `postselect_modes` argument from the argument list of the
`PostSelectPhotons` and `ImperfectPostSelectPhotons` classes;
* Updates relevant tests.