feat!: Branching logic for mid-circuit measurements
An improved logic for handling multiple measurements is added. This replaces
the original `_execute_multi_measurement_instructions`, which is executed when
multiple measurements are detected, and ran the entire circuit independently
`shots` many times. Instead, the measurement outcomes are binned and stored
with their frequency among all samples in a `Branch` object, along with the
corresponding post-measurement quantum state.

This means, that a "branching" is performed at a measurement, and the simulator
needs to update the list branches (instead of a single state) when applying
each instruction. By storing the frequencies of the branches, no two branch has
the same outcome, and hence, no redundant calculations are performed, yielding
a speedup over the previous multi-measurement logic. This logic also unifies
the single-measurement and the multi-measurement implementations. `Result`
offers the same functionality, but it stores `Branch` objects instead of just a
`State` object and a list of samples.

Breaking changes:
 - A calculation function (which has to be registered in `_instruction_map`)
   now needs to return with a list of `Branch` objects, instead of a single
   `Result` object. This is only relevant for a user using the API directly.
 - During exection, the list of instructions were modified previously for
   resolving `modes`, when it is not specified. With this change, the `modes`
   parameter for each instruction is left untouched.
 - The list of samples got a fixed type `List[Tuple[Union[int, float], ...]]`
   (previously one could get numpy arrays).