feat(simulator): Exact probabilities in `Result` using `shots=None`
The possibility of obtaining the exact probability distribution is enabled by
specifying `shots=None` at execution. This type of execution bypasses the
sampling stage, and instead returns with the full, exact probability
distribution, along with all possible outcomes with corresponding
post-measurement states. However, due to skipping sampling, `Result.samples`
has to be disabled in this case.

To keep track of the measurements which support this,
`Simulator._measurement_classes_allowed_with_shots_none` is created.