feat!(api): Runtime resolution of instruction parameters (#535)
The runtime resolution of instruction parameters is implemented in this
patch. This enables for users to specify callables (e.g., a lambda
function) as parameters of instructions, where the callable accepts a
tuple of outcomes and must return a float. This parameter is resolved
runtime during the program execution, after the program definition.