refact!(api): Runtime instruction validation
Instead of validating the instructions at program definition, the instructions
should only be validated during the simulation. This avoids issues when using a
connector other than `NumpyConnector`, because, e.g., using `JaxConnector`, the
instruction parameters are not regular numpy arrays, they might not have values
that can be accessed, and hence, the validation fails.

To overcome this, the method `BaseConnector.is_abstract` is added, which
defaults to returning `True`, and returns `False` for `JaxConnector` and
`TensorflowConnector` if the specified value is abstract (not concrete).
This method is practically use through `State._can_validate_variable`.

Moreover, the `State.validate()` methods were rewritten to validate
unconditionally to avoid the unexpected behaviour of not validating even if the
user is explicitely calling this method.

**Advantages** of this patch:
 - With this patch, we can avoid requiring the user to set `validate=False`
   when executing Piquasso code with non-eager JAX or TensorFlow variables.
   This makes using Piquasso more seamless.
 - It enables more differentiability in Piquasso while keeping some
   validations even in non-eager mode.
 - `State.validate()` has a more expected behaviour, i.e., it is not vacuous
   when `validate=False` is set.

**Disadvantages**:
 - Only runtime validation is possible, because both `config` and `connector`
   are needed to determine when to and what to validate.

**Breaking changes**
- `State.validate()` can raise an exception despite setting `validate=False` in
  `Config`.
- The instructions are validated during runtime instead of at program
  definition.