fix(tests): Fix seeded GBS test (#534)
The test `test_seeded_gaussian_boson_sampling_samples_displaced` has
been changed to be less brittle: instead of checking the samples against
hard-coded values, we check if two independent runs with the same seed
reproduces the same output. This is still sufficient, because we only
want to test the reproducability here.

Resolves #533